Extract miniprogram account client from setting
