+++ src/OpenPlatform/Authorizer/MiniProgram/Account/ServiceProvider.php
@@ -1,0 +1,25 @@
+<?php
+
+/*
+ * This file is part of the overtrue/wechat.
+ *
+ * (c) overtrue <i@overtrue.me>
+ *
+ * This source file is subject to the MIT license that is bundled
+ * with this source code in the file LICENSE.
+ */
+
+namespace EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Account;
+
+use Pimple\Container;
+use Pimple\ServiceProviderInterface;
+
+class ServiceProvider implements ServiceProviderInterface
+{
+    public function register(Container $app)
+    {
+        $app['account'] = function ($app) {
+            return new Client($app);
+        };
+    }
+}
