+++ src/OpenPlatform/Authorizer/MiniProgram/Application.php
@@ -22,6 +22,7 @@ use EasyWeChat\OpenPlatform\Authorizer\Aggregate\AggregateServiceProvider;
  * @property \EasyWeChat\OpenPlatform\Authorizer\Aggregate\Account\Client   $account
  * @property \EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Code\Client    $code
  * @property \EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Domain\Client  $domain
+ * @property \EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Account\Client  $setting
  * @property \EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Setting\Client  $setting
  * @property \EasyWeChat\OpenPlatform\Authorizer\MiniProgram\Tester\Client  $tester
  */
@@ -41,6 +42,7 @@ class Application extends MiniProgram
             AggregateServiceProvider::class,
             Code\ServiceProvider::class,
             Domain\ServiceProvider::class,
+            Account\ServiceProvider::class,
             Setting\ServiceProvider::class,
             Tester\ServiceProvider::class,
         ];
